+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Fixed missing blood pact death message
+
 ## 1.9.3
 * When spawners end being empowered now drop an enchanted book with a curse on it
 * Lodestone compasses can now be "reset" in a Grindstone